Fixed warnings in Map and Algorithms

+++ Raycaster-App/Source/Algorithms.h
@@ -22,7 +22,7 @@ public:
     static glm::u32vec2 AStar(glm::i32vec2 start, glm::i32vec2 end, std::vector<bool> map, uint32_t width, uint32_t height);
     static std::optional<glm::vec2> LineIntersection(glm::vec2 point1, glm::vec2 point2, glm::vec2 point3, glm::vec2 point4, bool isHalfLine = false);
     static glm::vec2 LineCollisions(glm::vec2 point, const std::vector<LineCollider>& lines, float thickness);
-    static std::vector<glm::vec2> MidpointCicle(glm::vec2 centre, float radius);
-    static std::vector<glm::ivec2> MidpointCicle(glm::ivec2 centre, int32_t radius);
+    static std::vector<glm::vec2> MidpointCicle(glm::vec2 centre, uint32_t radius);
+    static std::vector<glm::ivec2> MidpointCicle(glm::ivec2 centre, uint32_t radius);
 };
